pyproject.toml
[project]
name = "xtomarkdown"
version = "1.0.0"
description = "Cross-platform GUI application for converting documents to Markdown"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "Jesús Pérez", email = "tx2z@caosmental.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "PySide6>=6.6.0",
    "pypandoc-binary>=1.13",
    "markitdown[all]>=0.0.1",
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pyinstaller>=6.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[project.gui-scripts]
xtomarkdown = "xtomarkdown.app:main"

[project.scripts]
xtomarkdown-cli = "xtomarkdown.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/xtomarkdown"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/LICENSE",
    "/THIRD_PARTY_LICENSES",
    "/LICENSES",
    "/README.md",
]

# Ruff configuration (linting + formatting)
[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM108", # ternary operator (can reduce readability)
]

[tool.ruff.lint.isort]
known-first-party = ["xtomarkdown"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# Mypy configuration (type checking)
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "PySide6.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pypandoc.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "markitdown.*"
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

# Coverage configuration
[tool.coverage.run]
source = ["src/xtomarkdown"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
